🚚 refactor: 남은 shared 역방향 의존 제거 (Profile/Chip/hook 재배치)
[변경 내용]
- Profile, ProfileDropdown -> components/header로 이동 (member 정보 + 로그아웃
  + header 하위 컴포넌트를 조합하는 header 전용 위젯)
- ProjectStatusChip -> features/project/management/ui로 이동 (유일 소비자와 콜로케이트)
- useSchoolChapterMap -> entities/organization/hooks로 이동 (조직 도메인 훅)
- shared/api/me.ts 삭제 (소비자 0, entities/member/api/me의 getMyInfo와 중복된 죽은 코드)
- 소비자 import 경로 및 이동으로 깨진 상대 경로(ProfileDropdown의 shared/ui 기준
  상대 import) 절대 경로로 갱신

[변경 이유]
- Profile류는 member(entities) + auth 로그아웃(features) + header(components)를
  동시에 참조하는 3레이어 조합이라 shared에 두면 필연적으로 역방향. 세 레이어를
  모두 참조 가능한 유일한 위치가 components/header (components > features > entities)
- ProjectStatusChip은 project 도메인 전용이고 소비자도 project뿐이라 shared에 있을
  이유 없음. ProjectStatus 타입 재배치는 entities/project 단계로 분리
- useSchoolChapterMap은 organization api를 감싸는 조직 도메인 훅
- 이로써 shared/entities 레이어의 상위 역참조가 0이 됨 (테스트 파일 제외).
  FSD의 단방향 의존이 하위 레이어에서 완전히 회복
- boundaries 위반 95 -> 83건 (-12), shared 발 위반 전부 제거

[고려 사항]
- shared/api/me.ts 삭제 결정 근거: 소비자 grep 0건, 동일 기능(getMyInfo)이
  entities/member/api/me.ts에 이미 존재. 이동이 아닌 삭제가 맞다고 판단하여 명시
- shared/config/navigationResolve.test.ts의 components/sidebar 참조는 테스트
  파일이라 boundaries 검사 제외 대상 - 실제 위반 아님, 그대로 둠
- ProjectStatusChip을 관리 전용 폴더에 콜로케이트: 향후 재사용 필요 시 project
  공용 위치로 승격. 현재는 유일 소비자와 함께 두는 것이 결합도 최소
- 런타임 검증: tsc + 프로덕션 빌드 성공
- 남은 위반 83건은 전부 features 발(features->features 54, features->components
  26, components->components 3). entities/project 추출과 toast/tooltip의 shared
  이동으로 후속 정리 예정
