+++ src/components/header/Header.tsx
@@ -1,6 +1,7 @@
 import { Link, useLocation } from "@tanstack/react-router"
 import { useEffect, useState } from "react"
 
+import Profile from "@/components/header/Profile"
 import { MobileSidebarDrawerContent } from "@/components/sidebar/MobileSidebarDrawerContent"
 import { SideBarViewSwitcher } from "@/components/sidebar/SideBarViewSwitcher"
 import { useToastStore } from "@/components/toast/useToastStore"
@@ -8,7 +9,6 @@ import CloseIcon from "@/shared/assets/icon/close/CloseIcon"
 import HamburgerIcon from "@/shared/assets/icon/hamburger/HamburgerIcon"
 import UmcLogo from "@/shared/assets/icon/logo/UmcLogo"
 import { cn } from "@/shared/lib/utils"
-import Profile from "@/shared/ui/Profile"
 
 import HeaderButton from "./HeaderButton"
 import {
