+++ src/features/project/management/ui/ProjectManagementCard.tsx
@@ -3,9 +3,9 @@ import { useState } from "react"
 
 import { getProjectDetail } from "@/features/project/list/api/matchingProject"
 import { ProjectDetailCard } from "@/features/project/list/ui/ProjectDetailCard"
+import { ProjectStatusChip } from "@/features/project/management/ui/ProjectStatusChip"
 import { ProjectLogo } from "@/shared/assets/icon/logo/ProjectLogo"
 import { withImageCacheKey } from "@/shared/lib/withImageCacheKey"
-import { ProjectStatusChip } from "@/shared/ui/chip/ProjectStatusChip"
 import { RecruitStatusChip } from "@/shared/ui/chip/RecruitStatusChip"
 import MemberCount from "@/shared/ui/MemberCount"
 import { Modal } from "@/shared/ui/Modal"
