--- src/shared/api/me.ts
@@ -1,10 +1,0 @@
-import { api } from "@/shared/lib/axios"
-
-import type { MemberInfoResponse } from "@/entities/member/model/people"
-import type { ApiResponse } from "@/shared/lib/apiResponse"
-
-export async function getMe(): Promise<MemberInfoResponse> {
-  const { data } =
-    await api.get<ApiResponse<MemberInfoResponse>>("/v1/member/me")
-  return data.result
-}
